Debit left. Credit right.
Mistakenly thought for a long while now that credit meant to add or
deposit something to someone increasing their value, eg., the term
`credited to` means something positive, or that, `credit points` mean
a kind of an asset you can use to redeem something of value.

Apparently NOT.

English is weird. But I dont need to die on that hill. Not today.